refactor select input
