+++ src/components/form/FormContainer.tsx
@@ -4,6 +4,8 @@ type FormContainerProps = {
 
 export default function FormContainer(props: FormContainerProps) {
   return (
-    <form className="m-auto w-full md:w-3/5 p-4 pt-20">{props.children}</form>
+    <form className="m-auto w-full md:w-3/5 px-4 py-8 mt-10 bg-white rounded-md">
+      {props.children}
+    </form>
   );
 }
