+++ src/components/form/SelectBox/Select.tsx
@@ -1,15 +1,24 @@
 import ArrowDropDownIcon from "@mui/icons-material/ArrowDropDown";
 import { useState, useRef, useEffect } from "react";
 import { FormControlLabel } from "..";
+import SearchIcon from "@mui/icons-material/Search";
+import Chip from "@mui/material/Chip";
+import { Category } from "@data/api";
 
+type Option = {
+  value: string;
+  name: string;
+};
 type SelectProps = {
   startIcon?: React.ReactNode;
-  options?: { value: string; name: string }[];
+  options?: Option[];
   initValue?: { value: string; name: string };
   styles?: string;
   size?: string;
   label?: string;
   id?: string;
+  search?: boolean;
+  chip?: boolean;
 };
 
 export default function Select(props: SelectProps) {
@@ -18,14 +27,38 @@ export default function Select(props: SelectProps) {
     value: props.initValue?.value,
     name: props.initValue?.name,
   });
-
+  const [chosenSkills, setChosenSkills] = useState<Category[]>([]);
+  const [searchValue, setSearchValue] = useState<string>("");
+  const [filteredOptions, setFilteredOptions] = useState<Option[] | undefined>(
+    props.options
+  );
   const selectRef = useRef<HTMLDivElement>(null);
 
+  const changeHandler = (e: React.ChangeEvent<HTMLInputElement>) => {
+    setSearchValue(e.target.value);
+    const filtered = props.options?.filter((option) =>
+      option.name.toLowerCase().includes(e.target.value.toLowerCase())
+    );
+    setFilteredOptions(filtered || []);
+  };
+
   const onClickHandler = (e: React.MouseEvent<HTMLLIElement, MouseEvent>) => {
+    const selectedValue = e.currentTarget.getAttribute("value") || "";
+    const selectedName = e.currentTarget.textContent || "";
     setSelected({
       value: e.currentTarget.getAttribute("value") || "",
       name: e.currentTarget.textContent || "",
     });
+
+    const updatedChosenSkills = [
+      ...chosenSkills,
+      {
+        value: selectedValue,
+        name: selectedName,
+      },
+    ];
+
+    setChosenSkills(updatedChosenSkills);
     setOpen(false);
   };
 
@@ -57,13 +90,23 @@ export default function Select(props: SelectProps) {
         onClick={() => setOpen(!open)}
       >
         <div className="flex items-center">
-          <div>{props.startIcon}</div>
-          <span
-            className="text-gray-300 overflow-hidden whitespace-nowrap overflow-ellipsis"
-            id={selected.value}
-          >
-            {selected.name}
-          </span>
+          {props.startIcon && <div>{props.startIcon}</div>}
+          {!props.chip && (
+            <span
+              className="text-primary-600 font-semibold overflow-hidden whitespace-nowrap overflow-ellipsis"
+              id={selected.value}
+            >
+              {selected.name}
+            </span>
+          )}
+
+          {props.chip && (
+            <div className="text-gray-300 overflow-hidden whitespace-nowrap overflow-ellipsis">
+              {chosenSkills.map((skill) => (
+                <Chip label={skill.name} key={skill.value} />
+              ))}
+            </div>
+          )}
         </div>
 
         <div>
@@ -76,8 +119,23 @@ export default function Select(props: SelectProps) {
           open ? "absolute left-0 right-0 max-h-60 mt-2" : "max-h-0"
         }`}
       >
-        {props.options &&
-          props.options.map((option) => (
+        {props.search && (
+          <li className="relative p-1 items-center border-2 border-gray-150 bg-white">
+            <input
+              type="text"
+              className="pl-8 text-sm text-gray-300 border-none bg-transparent focus:outline-none"
+              placeholder="Tìm kiếm..."
+              onChange={changeHandler}
+              value={searchValue}
+            />
+            <div className="absolute inset-y-0 flex items-center">
+              <SearchIcon className="text-gray-300" />
+            </div>
+          </li>
+        )}
+        {filteredOptions &&
+          filteredOptions.length !== 0 &&
+          filteredOptions.map((option) => (
             <li
               className="p-2 text-sm hover:bg-primary-500 hover:text-white transition duration-100"
               key={option.value}
@@ -87,6 +145,16 @@ export default function Select(props: SelectProps) {
               {option.name}
             </li>
           ))}
+        {filteredOptions?.length === 0 && (
+          <li
+            className="p-2 text-sm hover:bg-primary-500 hover:text-white transition duration-100"
+            key="-1"
+            value="-1"
+            onClick={onClickHandler}
+          >
+            Không tìm thấy kết quả
+          </li>
+        )}
       </ul>
     </div>
   );
