+++ src/features/setting/components/CvProfile.tsx
@@ -1,4 +1,4 @@
-import { SearchSelect, Select, TextInput } from "@components/form";
+import { Select, TextInput } from "@components/form";
 
 import Typography from "@mui/material/Typography";
 import WorkIcon from "@mui/icons-material/Work";
@@ -12,7 +12,6 @@ import {
   majors,
   skills,
 } from "@data/api";
-import { SkillSelect } from "..";
 
 const genders = [
   {
@@ -63,22 +62,26 @@ export default function CvProfile() {
               id="position"
               type="text"
             ></TextInput>
-            <SearchSelect
+            <Select
+              search
               label="Chọn ngành nghề"
               initValue={majors[0]}
               options={majors}
               id="major"
             />
-            <SkillSelect
+            <Select
+              chip
+              search
               label="Chọn kỹ năng"
               initValue={skills[0]}
               options={skills}
             />
-            <SearchSelect
+            <Select
+              search
               initValue={locations[0]}
               options={locations}
               label="Chọn nơi làm việc"
-            ></SearchSelect>
+            ></Select>
             <Select
               initValue={expieriences[0]}
               options={expieriences}
