+++ src/features/setting/index.ts
@@ -3,4 +3,3 @@ export { default as SideBar } from "./components/SideBar";
 export { default as Profile } from "./components/Profile";
 export { default as CvProfile } from "./components/CvProfile";
 export { default as CvManage } from "./components/CvManage";
-export { default as SkillSelect } from "./components/SkillSelect";
