+++ src/pages/Register.tsx
@@ -1,3 +1,4 @@
+import { Footer } from "@components/layouts/footer";
 import { Header } from "@components/layouts/header";
 import { RegisterForm } from "@features/authentication";
 
@@ -6,6 +7,7 @@ export default function Register() {
     <>
       <Header></Header>
       <RegisterForm></RegisterForm>
+      <Footer></Footer>
     </>
   );
 }
